// src/app/Models/Pokemon.ts
export interface Pokemon
{
    id: number;
    name: string;
    hp: number;
    attack: number;
    // special?: number;
    defense: number;
    type1: string;
    // rarity?: number;
    // childID?: number;
    // parentID?: number;
    // combatAI?: string;
    // specialAttack?: string;

    // constructor (id: number, name: string, hp: number, attack: number, defense: number, type: string) {
    //     this.id = id;
    //     this.name = name;
    //     this.hp = hp;
    //     this.attack = attack;
    //     this.defense = defense;
    //     this.type = type;
    // }
}